+++ packages/backend/src/services/GdriveService/GdriveService.ts
@@ -8,6 +8,7 @@ import {
 } from '@lightdash/common';
 import { LightdashConfig } from '../../config/parseConfig';
 import { DashboardModel } from '../../models/DashboardModel/DashboardModel';
+import { ProjectModel } from '../../models/ProjectModel/ProjectModel';
 import { SavedChartModel } from '../../models/SavedChartModel';
 import { UserModel } from '../../models/UserModel';
 import { SchedulerClient } from '../../scheduler/SchedulerClient';
@@ -21,6 +22,7 @@ type GdriveServiceArguments = {
     dashboardModel: DashboardModel;
     userModel: UserModel;
     schedulerClient: SchedulerClient;
+    projectModel: ProjectModel;
 };
 
 export class GdriveService extends BaseService {
@@ -36,6 +38,8 @@ export class GdriveService extends BaseService {
 
     schedulerClient: SchedulerClient;
 
+    projectModel: ProjectModel;
+
     constructor({
         lightdashConfig,
         userModel,
@@ -43,6 +47,7 @@ export class GdriveService extends BaseService {
         savedChartModel,
         dashboardModel,
         schedulerClient,
+        projectModel,
     }: GdriveServiceArguments) {
         super();
         this.lightdashConfig = lightdashConfig;
@@ -51,18 +56,22 @@ export class GdriveService extends BaseService {
         this.savedChartModel = savedChartModel;
         this.dashboardModel = dashboardModel;
         this.schedulerClient = schedulerClient;
+        this.projectModel = projectModel;
     }
 
     async scheduleUploadGsheet(
         user: SessionUser,
         gsheetOptions: UploadMetricGsheet,
     ) {
+        const projectSummary = await this.projectModel.getSummary(
+            gsheetOptions.projectUuid,
+        );
         if (
             user.ability.cannot(
                 'manage',
                 subject('ExportCsv', {
-                    organizationUuid: user.organizationUuid,
-                    projectUuid: gsheetOptions.projectUuid,
+                    organizationUuid: projectSummary.organizationUuid,
+                    projectUuid: projectSummary.projectUuid,
                 }),
             )
         ) {
@@ -76,8 +85,8 @@ export class GdriveService extends BaseService {
             user.ability.cannot(
                 'manage',
                 subject('CustomSql', {
-                    organizationUuid: user.organizationUuid,
-                    projectUuid: gsheetOptions.projectUuid,
+                    organizationUuid: projectSummary.organizationUuid,
+                    projectUuid: projectSummary.projectUuid,
                 }),
             )
         ) {