+++ packages/backend/src/services/ServiceRepository.ts
@@ -342,6 +342,7 @@ export class ServiceRepository
                     dashboardModel: this.models.getDashboardModel(),
                     savedChartModel: this.models.getSavedChartModel(),
                     schedulerClient: this.clients.getSchedulerClient(),
+                    projectModel: this.models.getProjectModel(),
                 }),
         );
     }
@@ -804,6 +805,7 @@ export class ServiceRepository
                     lightdashConfig: this.context.lightdashConfig,
                     spotlightTableConfigModel:
                         this.models.getSpotlightTableConfigModel(),
+                    projectModel: this.models.getProjectModel(),
                 }),
         );
     }